Reorganize folders - Move sentence report files out of claim review folder - Move field list files for the dynamic form into a folder
