+++ src/components/ClaimReview/form/fieldLists/reportedForm.ts
@@ -1,7 +1,11 @@
 import { createFormField, FormField } from "./FormField";
 
 const reportedForm: FormField[] = [
-    createFormField({ fieldName: "classification", type: "select", defaultValue: "" }),
+    createFormField({
+        fieldName: "classification",
+        type: "select",
+        defaultValue: "",
+    }),
 ];
 
 export default reportedForm;
