+++ src/components/SentenceReport/Banner.style.tsx
