+++ src/components/SentenceReport/Banner.tsx
