+++ src/components/SentenceReport/SentenceReportCard.style.tsx
