+++ src/components/SentenceReport/SentenceReportCard.tsx
