+++ src/components/SentenceReport/SentenceReportContent.style.tsx
