+++ src/components/SentenceReport/SentenceReportContent.tsx
