+++ src/components/SentenceReport/SentenceReportView.style.tsx
