+++ src/components/SentenceReport/SentenceReportView.tsx
